refactor common link logic from link.py -> linkCommon
